Refactoring HTTP downlader progress reporter to accept multiple observers
